requirements.txt
google-generativeai
langchain
python-dotenv
langchain_experimental
pymysql
cryptography
streamlit
sentence-transformers
chormadb